Move worker to api and add app source file
