+++ dmci/api/__init__.py
@@ -1,6 +1,7 @@
+# -*- coding: utf-8 -*-
 """
-DMCI : Api init
-=================
+DMCI : API App Class
+====================
 
 Copyright 2021 MET Norway
 
@@ -16,54 +17,9 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 """
-import logging
-import os
-import uuid
 
-from flask import request, Flask, after_this_request
+from dmci.api.app import App
 
-from dmci import CONFIG
-from dmci.worker import Worker
-
-logger = logging.getLogger(__name__)
-
-class App():
-
-    def __init__(self):
-        self._app = Flask(__name__)
-        self._conf = CONFIG
-
-        @self._app.route("/v1/", methods=["POST"])
-        def base():
-            data = request.get_data()
-
-            file_uuid = uuid.uuid4()
-            path = self._conf.distributor_input_path
-            full_path = os.path.join(path, f"{file_uuid}.Q")
-
-            worker = Worker(full_path)
-
-            @after_this_request
-            def dist(response):
-                nonlocal worker
-                worker.distribute()
-                return response
-
-            result, msg = worker.validate(data)
-
-            if result:
-                return self._persist_file(data, full_path)
-            else:
-                return msg, 500
-
-        return
-
-    def _persist_file(self, data, full_path):
-        try:
-            with open(full_path, "wb") as queuefile:
-                queuefile.write(data)
-
-        except Exception as e:
-            logger.error(str(e))
-            return "Can't write to file", 507
-        return "", 200
+__all__ = [
+    "App",
+]
